job_gate_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
openai==1.6.1
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.4
pydantic==2.5.0
PyJWT==2.8.0
